tagging 1.3.12

* tag 'v1.3.12': (36 commits)
  bumping version
  Switch to homebrew install instructions
  Get default compile flags from homebrew
  Fix build on Ruby trunk
  remove hack for old versions of Rails.
  I noticed that the example in the README.rdoc has a small issue: It doesn't _select_ the inserted rows from the second table. I inserted an additional select statement into the example.
  no more 1.8
  Update rake-compiler-dock to support RubyInstaller-2.3.0
  clean up warnings from rake test
  SQLITE_VERSION_NUMBER #ifdefs surrounding sqlite3_open_v2 constant defines
  add 'readwrite' and 'flags' options to Database.new
  add travis build status
  fix Temfile usage on older Ruby
  test `filename` with attached databases
  move db_filename to be private, expose `filename` with default params
  add SQLite3::Database#db_filename
  add thread safety predicate to SQLite3
  Makes private type-based pragma helper methods public.
  Adds missing helper methods for pragmas.
  clean up test a little
  ...

+++ test/test_database_readonly.rb
@@ -3,7 +3,7 @@ require 'helper'
 module SQLite3
   class TestDatabaseReadonly < SQLite3::TestCase
     def setup
-      File.unlink 'test-readonly.db' if File.exists?('test-readonly.db')
+      File.unlink 'test-readonly.db' if File.exist?('test-readonly.db')
       @db = SQLite3::Database.new('test-readonly.db')
       @db.execute("CREATE TABLE foos (id integer)")
       @db.close
@@ -11,12 +11,19 @@ module SQLite3
 
     def teardown
       @db.close unless @db.closed?
-      File.unlink 'test-readonly.db'
+      File.unlink 'test-readonly.db' if File.exist?('test-readonly.db')
     end
 
     def test_open_readonly_database
       @db = SQLite3::Database.new('test-readonly.db', :readonly => true)
       assert @db.readonly?
+    end
+
+    def test_open_readonly_not_exists_database
+      File.unlink 'test-readonly.db'
+      assert_raise(SQLite3::CantOpenException) do
+        @db = SQLite3::Database.new('test-readonly.db', :readonly => true)
+      end
     end
 
     def test_insert_readonly_database
